worker/requirements.txt
jupygrader
psycopg2-binary
boto3
openai
